Tiare: Go: Disable memcg kernel and socket accounting
Disable memcg kernel and socket accounting on msm8937_32go since
memory cgroups is used in this target as a means of user space task
grouping based on oom_score_adj, and controlling the userspace
memory consumed by tasks. Kernel and socket memory accounting is
not considered. Since memcg is created per process on this target
enabling kernel accounting results in the creation of numerous
kmem caches resulting in significant overhead.

Change-Id: I3ca4758d9de9437651c45637f4310ba827d044ee